+++ DirectX11Engine/DirectX11Engine/Engine.h
@@ -6,7 +6,7 @@ private:
 	Graphics gfx;
 	RenderWindow m_window;
 public:
-	bool Initialize(HINSTANCE hInstance, const std::string& windowTitle, const std::string& windowClass, int width, int height);
+	bool Initialize(HINSTANCE hInstance, const std::string& windowTitle, const std::string& windowClass, float width, float height);
 	void Update();
 	void RenderFrame();
 	bool ProcessMessage();
